Renamed App Package to 'org.fossasia.susi.ai'
